add the need classes in the notification system and add more enums as needed
