+++ src/main/java/org/fcai/OrderNotificationModule/Enums/CategoryName.java
@@ -1,0 +1,17 @@
+package org.fcai.OrderNotificationModule.Enums;
+
+public enum CategoryName {
+    CLOTHES,
+    ELECTRONICS,
+    FOOD,
+    FURNITURE,
+    GROCERIES,
+    HEALTH,
+    HOME,
+    KITCHEN,
+    OFFICE,
+    OUTDOORS,
+    SPORTS,
+    TOYS,
+    OTHER
+}
