+++ src/main/java/org/fcai/OrderNotificationModule/Enums/NotificationChannel.java
@@ -1,0 +1,6 @@
+package org.fcai.OrderNotificationModule.Enums;
+
+public enum NotificationChannel {
+    SMS,
+    EMAIL,
+}
