+++ src/main/java/org/fcai/OrderNotificationModule/Enums/NotificationLanguage.java
@@ -1,0 +1,6 @@
+package org.fcai.OrderNotificationModule.Enums;
+
+public enum NotificationLanguage {
+    English,
+    Arabic
+}
