+++ src/main/java/org/fcai/OrderNotificationModule/Enums/OrderStatus.java
@@ -1,0 +1,7 @@
+package org.fcai.OrderNotificationModule.Enums;
+
+public enum OrderStatus {
+    OrderPlaced,
+    OrderShipped,
+    OrderCancelled,
+}
