+++ src/main/java/org/fcai/OrderNotificationModule/Exceptions/CategoryNotFoundException.java
@@ -1,0 +1,7 @@
+package org.fcai.OrderNotificationModule.Exceptions;
+
+public class CategoryNotFoundException extends RuntimeException {
+    public CategoryNotFoundException(int id) {
+        super("Category with id " + id + " does not exist.");
+    }
+}
