+++ src/main/java/org/fcai/OrderNotificationModule/Exceptions/OrderCancellationDurationException.java
@@ -1,0 +1,7 @@
+package org.fcai.OrderNotificationModule.Exceptions;
+
+public class OrderCancellationDurationException extends RuntimeException {
+    public OrderCancellationDurationException(String message) {
+        super(message);
+    }
+}
