+++ src/main/java/org/fcai/OrderNotificationModule/Exceptions/OrderNotFoundException.java
@@ -1,0 +1,7 @@
+package org.fcai.OrderNotificationModule.Exceptions;
+
+public class OrderNotFoundException extends RuntimeException {
+    public OrderNotFoundException(int id) {
+        super("Order with id = " + id + " not found.");
+    }
+}
