+++ src/main/java/org/fcai/OrderNotificationModule/Exceptions/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package org.fcai.OrderNotificationModule.Exceptions;
+
+public class ProductNotFoundException extends RuntimeException {
+    public ProductNotFoundException(int id) {
+        super("Product with id " + id + " does not exist.");
+    }
+}
