+++ src/main/java/org/fcai/OrderNotificationModule/Exceptions/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package org.fcai.OrderNotificationModule.Exceptions;
+
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException(int id) {
+        super("User with id " + id + " does not exist.");
+    }
+}
