+++ src/main/java/org/fcai/OrderNotificationModule/Models/Order.java
@@ -1,0 +1,9 @@
+package org.fcai.OrderNotificationModule.Models;
+
+import java.time.LocalDateTime;
+
+public interface Order {
+    double getTotalPrice();
+    String getOrderDetails();
+    LocalDateTime getCreationDate();
+}
