+++ src/main/java/org/fcai/OrderNotificationModule/OrderModuleApplication.java
@@ -1,0 +1,9 @@
+package org.fcai.OrderNotificationModule;
+
+import jakarta.ws.rs.ApplicationPath;
+import jakarta.ws.rs.core.Application;
+
+@ApplicationPath("/")
+public class OrderModuleApplication extends Application {
+
+}
