+++ src/main/java/org/fcai/OrderNotificationModule/Repositories/OrderRepository.java
@@ -1,0 +1,27 @@
+package org.fcai.OrderNotificationModule.Repositories;
+
+import org.fcai.OrderNotificationModule.Models.Order;
+
+import java.util.List;
+
+public class OrderRepository {
+    public List<Order> getAll() {
+        return null; // TODO: implement this method.
+    }
+
+    public Order getById(int id) {
+        return null; // TODO: implement this method.
+    }
+
+    public void createSimpleOrder(Order order) {
+        // TODO: implement this method.
+    }
+
+    public void createCompoundOrder(Order order) {
+        // TODO: implement this method.
+    }
+
+    public void delete(Order order) {
+        // TODO: implement this method.
+    }
+}
