+++ src/main/java/org/fcai/OrderNotificationModule/Repositories/ProductRepository.java
@@ -1,0 +1,38 @@
+package org.fcai.OrderNotificationModule.Repositories;
+import org.fcai.OrderNotificationModule.Models.Product;
+
+import java.util.List;
+
+public class ProductRepository {
+    public List<Product> getAll() {
+        return null; // TODO: implement this method.
+    }
+
+    public Product getById(int id) {
+        return null; // TODO: implement this method.
+    }
+
+    public void add(Product product) {
+        // TODO: implement this method.
+    }
+
+    public void remove(int id) {
+        // TODO: implement this method.
+    }
+
+    public List<Product> getByCategory(int categoryId) {
+        return null; // TODO: implement this method.
+    }
+
+    public void updateQuantity(int id, int quantity) {
+        // TODO: implement this method.
+    }
+
+    public void updatePrice(int id, double price) {
+        // TODO: implement this method.
+    }
+
+    public void updateCategory(int id, int categoryId) {
+        // TODO: implement this method.
+    }
+}
