+++ src/main/java/org/fcai/OrderNotificationModule/Repositories/UserRepository.java
@@ -1,0 +1,13 @@
+package org.fcai.OrderNotificationModule.Repositories;
+
+import org.fcai.OrderNotificationModule.Models.User;
+
+public class UserRepository {
+    public User isUserExist(String username, String password) {
+        return null; // TODO: Implement this method.
+    }
+
+    public void registerNewUser(User user) {
+        // TODO: Implement this method.
+    }
+}
