--- src/main/java/org/fcai/ordernotificationmodule/Exceptions/CategoryNotFoundException.java
@@ -1,7 +1,0 @@
-package org.fcai.ordernotificationmodule.Exceptions;
-
-public class CategoryNotFoundException extends RuntimeException {
-    public CategoryNotFoundException(int id) {
-        super("Category with id " + id + " does not exist.");
-    }
-}
