--- src/main/java/org/fcai/ordernotificationmodule/Exceptions/OrderCancellationDurationException.java
@@ -1,7 +1,0 @@
-package org.fcai.ordernotificationmodule.Exceptions;
-
-public class OrderCancellationDurationException extends RuntimeException {
-    public OrderCancellationDurationException(String message) {
-        super(message);
-    }
-}
