--- src/main/java/org/fcai/ordernotificationmodule/Exceptions/OrderNotFoundException.java
@@ -1,7 +1,0 @@
-package org.fcai.ordernotificationmodule.Exceptions;
-
-public class OrderNotFoundException extends RuntimeException {
-    public OrderNotFoundException(int id) {
-        super("Order with id = " + id + " not found.");
-    }
-}
