--- src/main/java/org/fcai/ordernotificationmodule/Exceptions/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package org.fcai.ordernotificationmodule.Exceptions;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(int id) {
-        super("Product with id " + id + " does not exist.");
-    }
-}
