--- src/main/java/org/fcai/ordernotificationmodule/Exceptions/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package org.fcai.ordernotificationmodule.Exceptions;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(int id) {
-        super("User with id " + id + " does not exist.");
-    }
-}
