--- src/main/java/org/fcai/ordernotificationmodule/Models/CategoryName.java
@@ -1,17 +1,0 @@
-package org.fcai.ordernotificationmodule.Models;
-
-public enum CategoryName {
-    CLOTHES,
-    ELECTRONICS,
-    FOOD,
-    FURNITURE,
-    GROCERIES,
-    HEALTH,
-    HOME,
-    KITCHEN,
-    OFFICE,
-    OUTDOORS,
-    SPORTS,
-    TOYS,
-    OTHER
-}
