--- src/main/java/org/fcai/ordernotificationmodule/Models/NotificationChannel.java
@@ -1,6 +1,0 @@
-package org.fcai.ordernotificationmodule.Models;
-
-public enum NotificationChannel {
-    SMS,
-    EMAIL,
-}
