--- src/main/java/org/fcai/ordernotificationmodule/Models/Order.java
@@ -1,9 +1,0 @@
-package org.fcai.ordernotificationmodule.Models;
-
-import java.time.LocalDateTime;
-
-public interface Order {
-    double getTotalPrice();
-    String getOrderDetails();
-    LocalDateTime getCreationDate();
-}
