--- src/main/java/org/fcai/ordernotificationmodule/OrderModuleApplication.java
@@ -1,9 +1,0 @@
-package org.fcai.ordernotificationmodule;
-
-import jakarta.ws.rs.ApplicationPath;
-import jakarta.ws.rs.core.Application;
-
-@ApplicationPath("/")
-public class OrderModuleApplication extends Application {
-
-}
