--- src/main/java/org/fcai/ordernotificationmodule/Repositories/OrderRepository.java
@@ -1,27 +1,0 @@
-package org.fcai.ordernotificationmodule.Repositories;
-
-import org.fcai.ordernotificationmodule.Models.Order;
-
-import java.util.List;
-
-public class OrderRepository {
-    public List<Order> getAll() {
-        return null; // TODO: implement this method.
-    }
-
-    public Order getById(int id) {
-        return null; // TODO: implement this method.
-    }
-
-    public void createSimpleOrder(Order order) {
-        // TODO: implement this method.
-    }
-
-    public void createCompoundOrder(Order order) {
-        // TODO: implement this method.
-    }
-
-    public void delete(Order order) {
-        // TODO: implement this method.
-    }
-}
