--- src/main/java/org/fcai/ordernotificationmodule/Repositories/UserRepository.java
@@ -1,13 +1,0 @@
-package org.fcai.ordernotificationmodule.Repositories;
-
-import org.fcai.ordernotificationmodule.Models.User;
-
-public class UserRepository {
-    public User isUserExist(String username, String password) {
-        return null; // TODO: Implement this method.
-    }
-
-    public void registerNewUser(User user) {
-        // TODO: Implement this method.
-    }
-}
